refactor(android): split MetricsScreen, sweep the governor's leftovers
MetricsScreen.kt had grown into a god-file: a CSV parser, file IO, a regex, the
Pearson math, a share intent and five views in 755 lines. The data layer moves to
MetricsCsv.kt (Csv, runTitle, fmt, pearson, shareCsv) and the screen keeps the
views — nothing in it touches a file now. 755 -> 655 + 125.

Governor-era leftovers, from a time when a governor moved the cache budget under
memory pressure:
  - compare defaulted to cache_budget_mib, "the column this whole feature was
    built to look at". The budget is fixed for a run now, so that default opened
    the view on a flat line. Defaults to cache_hit_pct.
  - correlate defaulted to {majflt_mib, rss_anon_mib, cache_budget_mib}; the third
    is now dense_resident_frac, which is what tells "the faults are the model"
    apart from "the faults are the cache". (Not cache_resident_mib: that lives in
    the summary trailer, not the per-token columns, so it would have silently
    fallen through to the first-two-columns fallback.)
  - Telemetry.cacheBudgetMib was documented as "possibly auto-adapting"; MetricFields
    already said "fixed for the run". Now they agree.
  - "was warm-dense on in both?" predates the DenseWeights enum: dense=warm.

Dead live-telemetry fields, parsed off the wire and never rendered: ioMs,
avgIoMs, denseResidentFrac. io_ms in particular is not an oversight — the panel
derives flash wait as wall - compute - mgmt on purpose, because under overlap the
read time is not a wall component (MainActivity: "instead of asking the user to
reconcile a compute residual with a parallel flash-I/O figure"). Dropping the
fields matches that decision.

Also: the n_predict default was literal 48 in both AppSettings and the service's
intent fallback (now AppSettings.DEFAULT_N_PREDICT), and the two charts each had
their own copy of the null-gap-skipping path builder (now DrawScope.drawSeries,
with the y-range choice left at the call site — that choice is what makes the two
charts different).

APK compiles (compilePlayDebugKotlin --rerun-tasks clean); UI-only, no engine change.
